+++ AdventOfCode/src/aoc19/Param.java
@@ -4,30 +4,30 @@ import java.util.function.Supplier;
 
 public class Param {
     private final ParamMode mode;
-    private final int val;
+    private final long val;
     private final RelativeBaseOffset relativeBaseOffset;
 
-    public Param(ParamMode mode, int val, RelativeBaseOffset relativeBaseOffset) {
+    public Param(ParamMode mode, long val, RelativeBaseOffset relativeBaseOffset) {
         this.mode = mode;
         this.val = val;
         this.relativeBaseOffset = relativeBaseOffset;
     }
 
-    public int getVal(int[] memory) {
+    public long getVal(long[] memory) {
         switch (mode) {
             case POSITION:
-                return memory[val];
+                return memory[(int) val];
             case IMMEDIATE:
                 return val;
             case RELATIVE:
-                return memory[val + this.relativeBaseOffset.getOffset()];
+                return memory[(int) (val + this.relativeBaseOffset.getOffset())];
             default:
                 throw new IllegalStateException("unexpected param mode: " + mode);
         }
     }
 
-    public void writeVal(int[] memory, int valToWrite) {
-        memory[val] = valToWrite;
+    public void writeVal(long[] memory, long valToWrite) {
+        memory[(int) val] = valToWrite;
     }
 
 }
